+++ ft_memcmp.c
@@ -6,7 +6,7 @@
 /*   By: ryishii <ryishii@student.42tokyo.j>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/20 17:22:10 by ryishii           #+#    #+#             */
-/*   Updated: 2020/11/27 00:39:18 by ryishii          ###   ########.fr       */
+/*   Updated: 2020/11/27 14:41:31 by ryishii          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,14 +14,14 @@
 
 int		ft_memcmp(const void *s1, const void *s2, size_t n)
 {
-	int				i;
+	size_t			i;
 	unsigned char	*a;
 	unsigned char	*b;
 
 	i = 0;
 	a = (unsigned char*)s1;
 	b = (unsigned char*)s2;
-	while (i < (int)n)
+	while (i < n)
 	{
 		if (a[i] != b[i])
 			return (a[i] - b[i]);
