+++ ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: ryishii <ryishii@student.42tokyo.j>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/02 23:15:04 by ryishii           #+#    #+#             */
-/*   Updated: 2020/11/27 00:41:38 by ryishii          ###   ########.fr       */
+/*   Updated: 2020/11/27 13:24:52 by ryishii          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,14 +16,14 @@ void	*ft_memcpy(void *dst, const void *src, size_t n)
 {
 	unsigned char	*s1;
 	unsigned char	*s2;
-	int				i;
+	size_t			i;
 
 	s1 = (unsigned char*)dst;
 	s2 = (unsigned char*)src;
 	i = 0;
 	if (dst == src)
 		return (dst);
-	while (i < (int)n)
+	while (i < n)
 	{
 		s1[i] = s2[i];
 		i++;
