+++ ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: ryishii <ryishii@student.42tokyo.j>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/09/16 00:26:18 by ryishii           #+#    #+#             */
-/*   Updated: 2020/11/26 04:16:47 by ryishii          ###   ########.fr       */
+/*   Updated: 2020/11/27 14:46:32 by ryishii          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int		ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int				i;
+	size_t			i;
 	unsigned char	*a;
 	unsigned char	*b;
 
@@ -23,13 +23,13 @@ int		ft_strncmp(const char *s1, const char *s2, size_t n)
 	i = 0;
 	if (n == 0)
 		return (0);
-	while (i < (int)n && a[i] == b[i])
+	while (i < n && a[i] == b[i])
 	{
 		if (s1[i] == '\0' || s2[i] == '\0')
 			return (0);
 		i++;
 	}
-	if (i == (int)n)
+	if (i == n)
 		return (0);
 	return (a[i] - b[i]);
 }
